requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
python-decouple==3.6
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
